Issue with button events on Edit page. Restructured home page.
