+++ app/src/main/java/com/example/kyliejue/bbafinalproject/EditActivity.java
@@ -1,0 +1,50 @@
+package com.example.kyliejue.bbafinalproject;
+
+import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.util.Log;
+import android.view.View;
+
+public class EditActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_edit);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        FloatingActionButton inTile = (FloatingActionButton) findViewById(R.id.in_tile);
+        inTile.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                setContentView(R.layout.activity_sensor); //FIXME
+                Log.d("STATE", "BUTTON PUSHED");
+            }
+        });
+
+        FloatingActionButton outTile = (FloatingActionButton) findViewById(R.id.out_tile);
+        outTile.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+//                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+//                        .setAction("Action", null).show();
+                Log.d("STATE","OUT BUTTON PUSHED");
+            }
+        });
+    }
+
+
+//    public void openOutTilePage() {
+//      setContentView(R.layout.activity_sensor); //FIXME
+//    }
+//
+//    public void openInTilePage() {
+//        //TODO
+//    }
+
+
+}
