+++ app/src/main/java/com/example/kyliejue/bbafinalproject/MainActivity.java
@@ -1,55 +1,19 @@
 package com.example.kyliejue.bbafinalproject;
 
-import android.hardware.SensorEventListener;
-import android.hardware.SensorManager;
-import android.hardware.Sensor;
-import android.hardware.SensorEvent;
-import android.os.Bundle;
-import android.app.Activity;
-import android.content.Context;
 
-public class MainActivity extends Activity implements SensorEventListener {
-    private SensorManager mSensorManager;
-    private Sensor mLight;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+
+public class MainActivity extends AppCompatActivity {
 
     @Override
-    public final void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        // Get an instance of the sensor service, and use that to get an instance of
-        // a particular sensor.
-        mSensorManager = (SensorManager) getSystemService(Context.SENSOR_SERVICE);
-        mLight = mSensorManager.getDefaultSensor(Sensor.TYPE_LIGHT);
     }
 
-    @Override
-    public final void onAccuracyChanged(Sensor sensor, int accuracy) {
-        // Do something here if sensor accuracy changes.
-    }
-
-    @Override
-    public final void onSensorChanged(SensorEvent event) {
-        float illuminance = event.values[0];
-        if (illuminance > 10.0) {
-            setContentView(R.layout.activity_sensor);
-        } else {
-            setContentView(R.layout.activity_main);
-        }
-        // Do something with this sensor data.
-    }
-
-    @Override
-    protected void onResume() {
-        // Register a listener for the sensor.
-        super.onResume();
-        mSensorManager.registerListener(this, mLight, SensorManager.SENSOR_DELAY_NORMAL);
-    }
-
-    @Override
-    protected void onPause() {
-        // Be sure to unregister the sensor when the activity pauses.
-        super.onPause();
-        mSensorManager.unregisterListener(this);
+    public void openEditPage(View view) {
+        setContentView(R.layout.activity_edit);
     }
 }
